recruitment-ui-service/requirements.txt
Flask==2.0.1
Flask-WTF==0.15.1
firefly-python==0.1.15
healthcheck==1.3.3

